backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Werkzeug
python-dotenv
psycopg2-binary
pydantic
email-validator